feat: rnnoise + multiple bug fixes
- mumble's rnnoise implementation, ReNameNoise is used
--- note: a DLL is needed in the bin folder for server executable to run.
--- DLL is included with commit
--- denoiser must operate on left and right channels separately for best results

- fix: "horror" voice effect
--- VoicePacketFromServer struct is 4 byte aligned
--- last 3 bytes of voice data were never transmitted

- fix: server processing causing packet loss
--- cause was imprecise sleep timing during Sleep
--- solved by using timeSetEvent instead of Sleep

+++ headers/common/jitter_buffer.h
@@ -13,7 +13,7 @@ struct VoicePacketToServer {
 
 struct VoicePacketFromServer {
   uint32_t packet_number;
-  uint8_t userID;
+  uint32_t userID;
   uint8_t encoded_data[SLOT_SIZE];
 };
